Per discussion with Philippe:
I am commiting this because I am leaving for vacation.  The changes in here remove the option of transcoding a wav to mp3 (why?) and 
force transcoding of mp3s to wavs and transcoding of wavs to acceptable wavs.  As discussed with Philippe, this requires sox and 
mpg123 to be installed, so there should be a amportal.conf configuration for disabling this functionality if the user doesn't have or 
doesn't want to install mpg123 (should be rare).